+++ packages/be-ts-db-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [4.2.9] (2026-02-22)
+
+### 🏗️ Changed
+
+- Updated dependencies: now requires `@aneuhold/be-ts-lib@^3.0.19`, `@aneuhold/core-ts-db-lib@^4.1.8`, and `@aneuhold/core-ts-lib@^2.4.0`.
+
 ## 🔖 [4.2.8] (2026-02-21)
 
 ### 🏗️ Changed
@@ -267,6 +273,7 @@ Updated dependencies: now requires `@aneuhold/core-ts-db-lib@^3.0.0`, `@aneuhold
 
 <!-- Link References -->
 
+[4.2.9]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.8...be-ts-db-lib-v4.2.9
 [4.2.8]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.7...be-ts-db-lib-v4.2.8
 [4.2.7]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.6...be-ts-db-lib-v4.2.7
 [4.2.6]: https://github.com/aneuhold/ts-libs/compare/be-ts-db-lib-v4.2.5...be-ts-db-lib-v4.2.6