+++ packages/core-ts-lib/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [2.4.0] (2026-02-22)
+
+### ✅ Added
+
+- Added `DateService.getDaySinceEpoch()` static method to get the calendar day number since the Unix epoch for a given date in local time.
+- Added `DateService.getCalendarDaysBetween()` static method to get the number of calendar days between two dates, ignoring the time-of-day component.
+
 ## 🔖 [2.3.19] (2026-02-21)
 
 ### 🏗️ Changed
@@ -196,6 +203,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Link References -->
 
+[2.4.0]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.19...core-ts-lib-v2.4.0
 [2.3.19]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.18...core-ts-lib-v2.3.19
 [2.3.18]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.17...core-ts-lib-v2.3.18
 [2.3.17]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.16...core-ts-lib-v2.3.17